+++ lib/maplibre-compose/src/commonMain/kotlin/dev/sargunv/maplibrecompose/compose/source/rememberComputedSource.kt
@@ -1,6 +1,7 @@
 package dev.sargunv.maplibrecompose.compose.source
 
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.key
 import dev.sargunv.maplibrecompose.core.source.ComputedSource
 import dev.sargunv.maplibrecompose.core.source.ComputedSourceOptions
 import io.github.dellisd.spatialk.geojson.BoundingBox
@@ -14,7 +15,9 @@ public fun rememberGeoJsonSource(
   options: ComputedSourceOptions = ComputedSourceOptions(),
   getFeatures: (bounds: BoundingBox, zoomLevel: Int) -> FeatureCollection,
 ): ComputedSource =
-  rememberUserSource(
-    factory = { ComputedSource(id = it, options = options, getFeatures = getFeatures) },
-    update = {},
-  )
+  key(options, getFeatures) {
+    rememberUserSource(
+      factory = { ComputedSource(id = it, options = options, getFeatures = getFeatures) },
+      update = {},
+    )
+  }
