+++ lib/maplibre-compose/src/commonMain/kotlin/dev/sargunv/maplibrecompose/compose/source/rememberRasterSource.kt
@@ -1,6 +1,7 @@
 package dev.sargunv.maplibrecompose.compose.source
 
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.key
 import dev.sargunv.maplibrecompose.core.source.Defaults
 import dev.sargunv.maplibrecompose.core.source.RasterSource
 import dev.sargunv.maplibrecompose.core.source.TileSetOptions
@@ -11,10 +12,12 @@ public fun rememberRasterSource(
   uri: String,
   tileSize: Int = Defaults.RASTER_TILE_SIZE,
 ): RasterSource =
-  rememberUserSource(
-    factory = { RasterSource(id = it, uri = uri, tileSize = tileSize) },
-    update = {},
-  )
+  key(uri, tileSize) {
+    rememberUserSource(
+      factory = { RasterSource(id = it, uri = uri, tileSize = tileSize) },
+      update = {},
+    )
+  }
 
 @Composable
 public fun rememberRasterSource(
@@ -22,7 +25,9 @@ public fun rememberRasterSource(
   options: TileSetOptions = TileSetOptions(),
   tileSize: Int = Defaults.RASTER_TILE_SIZE,
 ): RasterSource =
-  rememberUserSource(
-    factory = { RasterSource(id = it, tiles = tiles, options = options, tileSize = tileSize) },
-    update = {},
-  )
+  key(tiles, options, tileSize) {
+    rememberUserSource(
+      factory = { RasterSource(id = it, tiles = tiles, options = options, tileSize = tileSize) },
+      update = {},
+    )
+  }
