+++ lib/maplibre-compose/src/commonMain/kotlin/dev/sargunv/maplibrecompose/compose/source/rememberUserSource.kt
@@ -1,7 +1,6 @@
 package dev.sargunv.maplibrecompose.compose.source
 
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.DisposableEffect
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.remember
 import dev.sargunv.maplibrecompose.compose.engine.LocalStyleNode
@@ -10,18 +9,9 @@ import dev.sargunv.maplibrecompose.core.source.Source
 @Composable
 internal fun <T : Source> rememberUserSource(factory: (String) -> T, update: T.() -> Unit): T {
   val node = LocalStyleNode.current
-  val source = remember(factory, node) { factory(node.sourceManager.nextId()) }
+  val source = remember(node) { factory(node.sourceManager.nextId()) }
   LaunchedEffect(source, update, node.style.isUnloaded) {
     if (!node.style.isUnloaded) source.update()
   }
   return source
-}
-
-@Composable
-internal fun SourceReferenceEffect(source: Source) {
-  val node = LocalStyleNode.current
-  DisposableEffect(source) {
-    node.sourceManager.addReference(source)
-    onDispose { node.sourceManager.removeReference(source) }
-  }
 }
